JSX

Documentation for JSX.

JSX.JSX

JSX.Node

JSX.findlimit

JSX.hasfinished

JSX.isblank

JSX.literalorquote

JSX.parse

JSX.parse!

JSX.processattr

JSX.pushexprorstr!

JSX.pushorappend!

JSX.toexpr

JSX.toexpr

JSX.toexpr

JSX.toexpr

JSX.trycatchexpr

JSX.vec2expr

JSX.@htm_str

JSX.JSX — Module

HTML parsing on steroids!

It's basically a tree traversal!

JSX.Node — Type

Node{T}

A node in the tree.

JSX.findlimit — Function

Find the position where the content ends.

JSX.hasfinished — Function

Check whether we have reached the end of the current tag.

JSX.isblank — Method

Check if a string is empty of pure whitespace.

JSX.literalorquote — Method

Attempt to parse a literal value, or quote and return a string.

JSX.parse! — Function

Parse a tree by modifying the a dummy node.

JSX.parse — Method

Parse a tree.

JSX.processattr — Method

Process and normalize attributes.

JSX.pushexprorstr! — Method

Push strings only if the last object in the vector is not a string, concatenate otherwise. This ensures strings are as long as possible.

We also ensure the "$(Expr(:incomplete, "incomplete: invalid string syntax"))

JSX.pushorappend! — Method

Append if vector, push otherwise.

JSX.toexpr — Method

Create a generic expression for a branch node.

JSX.toexpr — Method

Create a generic expression for a component node.

JSX.toexpr — Method

Create a generic expression for a leaf node.

JSX.toexpr — Method

Construct an expression that evaluates to the given node.

JSX.trycatchexpr — Method

Wrap an expression in a try...catch block.

JSX.vec2expr — Method

Recreate a vector as an expression. Optionally, apply a function to each element beforehand.

JSX.@htm_str — Macro

Time to create a macro!

We want to do the following.

Examples

julia> htm"<h1>Hello world!</h1>"
<h1>Hello world&#33;</h1>