Docstrings

HTM.Tag — Type

Tag(type, props[, promises=(), children=()])

Compile time internal representation of an HTML tag.

Base.readuntil — Method

readuntil(predicate, io::IO)

Read characters until matching a predicate.

Based on https://github.com/JuliaLang/julia/issues/21355#issue-221121166.

Base.startswith — Method

startswith(io::IO, prefix::Union{AbstractString,Base.Chars})

Check if an IO object starts with a prefix.

Based on https://github.com/JuliaLang/julia/issues/40616#issue-867861851.

HTM.create_element — Method

create_element(type[, children...])
create_element(type, props[, children...])

Create a Hyperscript.jl element.

This is an alternative syntax and (currently) serves as a rather trivial absctraction layer inspired by React.createElement.

HTM.parse — Method

parse(io::IO)
parse(s::AbstractString)

Parse HTML.

HTM.parseelem — Method

parseelem(io::IO)

Parse a single HTML element.

HTM.parseelems — Method

parseelems(io::IO)

Parse HTML elements.

This function is the entry point for an implementation of a subset of the HTML standard.

HTM.parsekey — Method

parsekey(io::IO)

Parse an HTML property key.

HTM.parseprops — Method

parseprops(io::IO)

Parse HTML properties of a tag.

HTM.parsetag — Method

parsetag(io::IO)

Parse a Tag object.

HTM.parsetagtype — Method

parsetagtype(io::IO)

Parse an HTML tag type.

HTM.parsevalue — Method

parsevalue(io::IO)

Parse an HTML property value.

HTM.skipstartswith — Method

skipstartswith(io::IO, prefix::Union{AbstractString,Base.Chars})

Check if an IO object starts with a prefix and skip it.

Index

HTM.Tag

Base.readuntil

Base.startswith

HTM.create_element

HTM.parse

HTM.parseelem

HTM.parseelems

HTM.parsekey

HTM.parseprops

HTM.parsetag

HTM.parsetagtype

HTM.parsevalue

HTM.skipstartswith