Docstrings

HTM.Node — Type

Node(tag, attrs, promises=String[], children=Union{String, HTM.Node}[])

Compile time internal representation of an HTML node.

julia> HTM.Node(["div"], Dict("class" => ["fruit"]), [], ["🍍"])
HTM.Node(["div"], Dict("class" => ["fruit"]), String[], Union{String, HTM.Node}["🍍"])

Base.readuntil — Method

readuntil(predicate, io::IO)

Read characters until matching a predicate.

Based on this.

julia> readuntil(isspace, IOBuffer("pineapple: 🍍..."))
"pineapple:"

Base.startswith — Method

startswith(io::IO, prefix::Union{AbstractString,Base.Chars})

Check if an IO object starts with a prefix.

Based on this.

julia> io = IOBuffer("pineapple: 🍍...");

julia> startswith(io, "pine")
true

julia> startswith(io, "apple")
false

HTM.create_element — Method

create_element(tag, attrs[, children...])

Create a Hyperscript.jl element.

This is an alternative syntax and (currently) serves as a rather trivial absctraction layer inspired by React.createElement.

julia> create_element("div", Dict("class" => "fruit"), "🍍")
<div class="fruit">🍍</div>

HTM.parse — Method

parse(s::AbstractString)
parse(io::IO)

Parse HTML.

julia> HTM.parse("pineapple: <div class=\"fruit\">🍍</div>...")
3-element Vector{Union{String, HTM.Node}}:
 "pineapple: "
 HTM.Node(["div"], Dict("class" => ["fruit"]), String[], Union{String, HTM.Node}["🍍"])
 "..."

HTM.parseattrs — Method

parseattrs(io::IO)

Parse HTML attributes of a node. The returned tuple contains both true attributes and promisses.

julia> HTM.parseattrs(IOBuffer("class=\"fruit\">🍍..."))
(Dict("class" => ["fruit"]), String[])

julia> HTM.parseattrs(IOBuffer("class=\"fruit\" \$(attrs)>🍍..."))
(Dict("class" => ["fruit"]), ["\$(attrs)"])

HTM.parseelem — Method

parseelem(io::IO)

Parse a single HTML element.

julia> HTM.parseelem(IOBuffer("pineapple: <div class=\"fruit\">🍍</div>..."))
"pineapple: "

HTM.parseelems — Method

parseelems(io::IO)

Parse HTML elements.

This function is the entry point for an implementation of a subset of the HTML standard.

julia> HTM.parseelems(IOBuffer("pineapple: <div class=\"fruit\">🍍</div>..."))
3-element Vector{Union{String, HTM.Node}}:
 "pineapple: "
 HTM.Node(["div"], Dict("class" => ["fruit"]), String[], Union{String, HTM.Node}["🍍"])
 "..."

HTM.parseinterp — Method

parseinterp(io::IO)
parseinterp(fallback, io::IO)

Parse an interpolation as string, including $.

The input must start with $ if no fallback function is given. The fallback function is passed to readuntil if the input does not start with $.

julia> HTM.parseinterp(IOBuffer(raw"$((1, (2, 3)))..."))
"\$((1, (2, 3)))"

HTM.parsekey — Method

parsekey(io::IO)

Parse an HTML attribute key.

julia> HTM.parsekey(IOBuffer("class=\"fruit\">🍍..."))
"class"

HTM.parsenode — Method

parsenode(io::IO)

Parse a Node object.

julia> HTM.parsenode(IOBuffer("<div class=\"fruit\">🍍</div>..."))
HTM.Node(["div"], Dict("class" => ["fruit"]), String[], Union{String, HTM.Node}["🍍"])

HTM.parsetag — Method

parsetag(io::IO)

Parse an HTML tag.

julia> HTM.parsetag(IOBuffer("div class=\"fruit\">🍍..."))
1-element Vector{String}:
 "div"

HTM.parsevalue — Method

parsevalue(io::IO)

Parse an HTML attribute value.

julia> HTM.parsevalue(IOBuffer("\"fruit\">🍍..."))
1-element Vector{String}:
 "fruit"

HTM.skipstartswith — Method

skipstartswith(io::IO, prefix::Union{AbstractString,Base.Chars})

Check if an IO object starts with a prefix and skip it.

julia> io = IOBuffer("pineapple: 🍍...");

julia> HTM.skipstartswith(io, "pine")
true

julia> read(io, String)
"apple: 🍍..."

julia> io = IOBuffer("pineapple: 🍍...");

julia> HTM.skipstartswith(io, "apple")
false

julia> read(io, String)
"pineapple: 🍍..."

Index

HTM.Node

Base.readuntil

Base.startswith

HTM.create_element

HTM.parse

HTM.parseattrs

HTM.parseelem

HTM.parseelems

HTM.parseinterp

HTM.parsekey

HTM.parsenode

HTM.parsetag

HTM.parsevalue

HTM.skipstartswith